Initialize test suite
Start a test suite for the project to validate the output of the
scanner. The testing framework that will be used is AVA [1] (added as a
devDependency (which won't be installed in the Docker image due to line
22 of the Dockerfile)) and the approach will be to use snapshots of the
scanner output for testing. Per AVA's snapshot testing, these can be
found, both human-readable and binary, in tests/snapshots.

For this initial version of the test, there's just one test that runs
the scanner against the newly added testdata/ directory (containing a
single file with a bad regular expression, courtesy of shescape [2]).

In order to run the test one has to:

1. Install Node.js dependencies (`npm install`)
2. Build the Docker image
3. Run the tests using `npm run ava`

These steps are also performed in the CI pipelines to continuously run
the tests and ensure the scanner works (more or less) as expected.

The human-readable snapshots are included to make it easier to see how
the snapshots evolve over time and, in particular, tell what changed
when the snapshot is updated.

Lastly, the binary snapshots are explicitly marked as binary (not
diffable) in the .gitattributes file for convenience.

--
1. https://github.com/avajs/ava
2. https://github.com/ericcornelissen/shescape