Test `.mjs` and `.cjs` files
Add testdata for JavaScript files with the ESModules extension (`.mjs`)
and the CommonJS extension (`.cjs`) to ensure these are scanned.

Both testdata files contain a problematic regular expression (courtesy
of, again, shescape) that should be flagged by this scanner.

The snapshot has been updated to reflect these two added files. Note
that no test needs to be added because we scan the entire testdir/
folder in the sample test.