Test scanner by running it on shescape at v1.5.9
Update the tests to run on individual directories in the testdata/
folder. The pre-existing test samples have been moved to a subfolder of
testdata/, and the corresponding test has been updated to scan only that
subfolder.

Then, a second test case is added through embedding the shescape project
as a submodule into the testdata/ folder. Specifically, the submodule
points to v1.5.9 of shescape because it's the latest known version with
problematic regular expressions (all expressions were fixed in v1.5.10).

The snapshots were updated accordingly. Also, the "ava" command now runs
only the tests in the tests/ folder; this change is necessary because
otherwise it will try to run the tests of shescape as well.