Initialize the changelog
Start a changelog (`CHANGELOG.md`) to log changes to the project over
time. The format is modelled after the "Keep a Changelog" project (link
in `CHANGELOG.md`). It includes all changes that have happened up to
now. Each item has a reference to the commit that contains the relevant
change, allowing readers to find out more if they want.