Test scanner on a TypeScript project
Create a new sample test for running the scanner on a TypeScript (ts)
project. This test is added prior to support being added to the scanner,
so this test is expected to fail. The snapshot for this test is based
on the fact that:

- The faulty Regular Expression in the `.ts` file is take from the
  `sample.cjs` file.
- The call to ESLint must include the value `.ts` for the extension
  (`--ext`) option. (Putting it in the end is an arbitrary choice.)

The sample project intentionally includes some TypeScript-specific
syntax (namely the explicit type annotation `: RegExp`) to ensure
parsing the file doesn't accidentally work with the standard JavaScript
parser.