Add ESLint parser for TypeScript files
Add `typescript-eslint/parser` as a dependency to be able to properly
parse TypeScript files (`.ts`) and thus report on problematic regular
expressions within.

In particular, this is achieved by adding an overriding configuration
for ESLint for files that end with `.ts` ([^1]) to use the TypeScript
parser.

--
[^1]: The first pattern matches `.ts` files in the root, the second
pattern matches `.ts` files in the rest of the project. If you can
manage to achieve this with a single pattern, feel free to submit a
patch.