Initialize Dependabot configuration
Initialize a configuration file for Dependabot to keep Docker, GitHub
Actions, and npm dependencies up-to-date.

Note that git submodules are intentionally not included in the config.
The usage of git submodules in this project is to have a lightweight way
to import real-world projects with problematic regular expressions.
Since these projects will continue developing past the affected version,
we don't want to receive Pull Requests to update these projects to a
more recent (likely unaffected or affected differently) version.