Pin Node.js and Alpine version for the Docker image
Update the Dockerfile to "pin" the Node.js and Alpine Linux versions for
the image. Rather than just using _the latest_ Node.js v18 and Alpine
image - whatever that happens to be at build time - this ensures that we
always use (exactly) Node.js v18.9.0 and Alpine Linux 3.15 when the
image is build.

This is useful for reproducibility (the image doesn't change because the
latest Node.js 18 release changes since the last time we build) and
security (since we can say for certain which version of Node.js and
Alpine are used in the image, we can determine if those version have
vulnerabilities).

Note the quotation marks around "pin": the tag is still not 100% static,
it's just that the maintainers of the image promise to deliver Node.js
18.9.0 with Alpine Linux 3.15 in this image.