Pin GitHub Actions to SHAs
Update the GitHub Actions workflows to pin all GitHub Actions used to
the SHA of the commit corresponding to the version tag they were
previously referencing.

This is preferred as what a git tag points to _may_ change, whereas the
SHA of a commit is more difficult to replicate with different repository
contents.

Using SHAs in this way is supported by Dependabot.