Add more ignored directories based on GitHub's Node.gitignore
Update the list of `ignorePatterns` for ESLint with patterns coming from
GitHub's Node.gitignore [1]. An attempt was made to extract only
directories from the template.

Including these patterns in the ignore list for ESLint should improve
the usefulness of the scanner as it should result in less noise. There's
generally no point to be alerted about problematic regular expression in
generated or otherwise temporary code. Also, reporting on problematic
regular expressions in dependencies is likely to be too noisy (also, as
node_modules was already ignored, ignoring other dependency directories
only makes sense).

--
1. Referenced at commit e5323759e387ba347a9d50f8b0ddd16502eb71d4