Audit dependencies nightly
Update the CI configuration to audit dependencies nightly in addition to
on every push and Pull Request. This should provide better coverage, in
particular when the project doesn't receive changes on a particular day
(or longer).

This is achieved by using reusable workflows. The `reusable-audit.yml`
workflow is configured to be callable by other workflows. In particular,
it is called by the pre-existing `test.yml` workflow (meaning it works
(roughly) the same as before); In addition it's called from a new GitHub
Actions workflow called `nightly.yml` which is configured to run every
night, only to run the audit workflow.

The nightly workflow can also be triggered manually. This was added to
be able to quickly verify that the workflow accepts a fix for a problem
it detected.