Replace Snyk scan with Syft & Grype
Replace the Snyk scan in the reusable-audit.yml workflow by Syft and
Grype for SBOM creation and vulnerability scanning (resp.). The main
motivation being that Syft and Grype are open source (in contrast to
Snyk).

The current CI job is meant to create an SBOM (software bill of
materials) and scan that for vulnerabilities, and upload that as an
artifact for manual review. The job, as is, won't fail if known
vulnerabilities were detected.