Configure Grype to error when a vulnerability is detected
Update the Grype configuration to exit with a non-zero exit code when a
vulnerability is detected (excluding ignored vulns). This makes it
easier to tell if a vuln worth your attention was detected, and also
ensures the CI fails if a vuln was found.

Since the CI workflow will now fail if a vulnerability was found, the
step to upload the SBOM and vulnerabilities as an artifact was updated
to `${{ always() }}` run. This ensures the artifact is uploaded no
matter if a vulnerability was detected.

Also, add the output file and format for Grype to the configuration
file. Like for Syft, this means that file is normalized and can be
ignored by git. For that reason, both the Syft and Grype output files
are added to the .gitignore file. Similarly, the audit workflow has been
updated to just use the configured Grype output file/format.