Add auditing section to the Contributing Guidelines
Update the Contributing Guidelines to include an auditing section with
instructions on how to run Syft and Grype locally.

As the instructions for auditing the project rely on downloading Syft
and Grype into the project at `./bin/syft` and `./bin/grype` (resp.), a
.gitignore file was created at `./bin/.gitignore` to ensure 1) that the
directory exists and can be downloaded into, and 2) that the Syft and
Grype binaries (or any other binary, really) are never committed.