Only audit production dependencies for version refs
Update the reusable-audit.yml workflow to only audit production
dependencies with npm if the provided ref is for a version (tag or
branch).

This is to avoid warnings about development dependencies that may have
already been resolved (e.g. on main) but not yet released. As dev
dependencies don't affect the released scanner, warnings about them are
pointless at best (and noisy at worst).

This was implemented within the reusable workflow to reduce changes and
avoid introducing similar noise if the workflow is used in a different
way in the future.