Initialize a Makefile for the project
Create a Makefile with commands that aid in development of this project.
The available commands roughly correspond to command listed in the
Contributing Guidelines (which is to be updated). It also includes
dependencies between commands to ensure it's always possible to run any
particular command.

The package.json is updated such that scripts correspond to commands
available through npm. Having these available through the package.json
ensures that, when using `npm run xxx`, the correct binary is invoked
regardless of OS and npm version used when running `npm install`.