Temporarily create Syft and Grype config files for v0 branch
The configuration files for Syft and Grype don't exist yet at the v0
branch as of this commit. As a result, the nightly Docker audit for v0
is failing. To avoid that, the configuration files will be created
during the workflow for now.

This should be removed once v0 points to a commit with the appropriate
configuration files.