Use file-based rules to improve Makefile performance
Rewrite the Makefile to use rules based in files so that Make knows when
things have to be re-executed (or not). This prevents things such as
regenerating the SBOM if the Docker image didn't change.

There's a couple of potentially noteworthy approaches here:

- In some cases there's a phony synonym for one specific file (e.g.
  `audit-docker: | vulns.json`). This is to provide a convenient name
  for generating that file while at the same type allowing that file to
  be referenced in other rules.
- The build command depends on the file `.tmp/dockerimage`. This file is
  simply generated whenever the Docker image is build to use as a stand
  in reference to compare against the Dockerfile to see if the image has
  to be rebuild.

Lastly, the continuous integration workflow files have been updated to
use Make commands instead of custom commands as much as possible.