Add dependency-review-action to continuous integration
Update the dependency-review-action to the continuous integration set
up. This action scans the dependencies for vulnerabilities (not really
wanted, but okay) and license. The latter is what it's being added for.

The intent is to use this action to block any dependencies being added
with unknown licenses. Then, the license should be reviewed and either
allowed or, if the license prevents use of the dependency, the
dependency should be replaced.

The configuration for the action is added separately. It contains a
preliminary list of allowed licenses (based on direct npm dependencies
and this project's license) as well as configuration to scan as broadly
as possible.

The action is set up to always review all dependencies by comparing the
"current" commit against the first commit in the (same) history.