Add Fossa to continuous integration (using CLI)
Update Fossa to the continuous integration set up using the Fossa CLI.
Fossa can scan dependencies for licenses and enforce a policy on those
licenses (the policy is configured in the Fossa webapp).

The intent is to use this action to block any dependencies being added
with unknown or unwanted licenses. For unknown licenses, they should be
reviewed and either allowed or, if the license prevents use of the
dependency, the dependency should be replaced.

The Fossa configuration file is included primarily to prevent Fossa from
scanning projects included as submodules in the testdata/.

The CI is set up to run `fossa analyze`, which analyzes the project,
finds licenses, and uploads that the Fossa first. And then run `fossa
test`, which checks if any license is present that shouldn't be, and
fails/succeeds accordingly.

These commands can be run locally using the Fossa CLI to obtain the same
result; This does require an API token, so if you do it with your own
token the result will be uploaded to your account.