Add licensed to continuous integration
Add Licensed to the continuous integration set up using the action
jonabc/setup-licensed to install Licensed and invoke it to catalogue all
licenses (npm dependencies only), create a NOTICE file (uploaded as an
artifact), and check for any disallowed/unknown licenses.

The intent is to use this action to block any dependencies being added
with unknown or unwanted licenses. For unknown licenses, they should be
reviewed and either allowed or, if the license prevents use of the
dependency, the dependency should be replaced.

A configuration file for licensed is included. This includes a list of
licenses that are allowed. The list comes from licenses in current
dependencies. It also includes a list of "reviewed" dependencies, for
all of these Licensed wasn't able to determine the license of the
dependency, so I reviewed their licenses manually and added them as a
comment for visibility.[^1]

The `status` command is the one that checks for unwanted/unknown
licenses. If any are detected, it exits with a non-zero exit code,
hence it must be executed after uploading the NOTICE, otherwise the
NOTICE would be missing only because of a potentially unwanted license.

--
[^1]: using the `confidence_threshold` option did not seem to have an
effect on the results I got locally. So I ended up not even including it
in the configuration.