Add Make commands to install and run licensed
Update the Makefile with two user-facing commands:

- `make license-check-npm`: to check the licenses of npm dependencies
  (using Licensed).
- `make notice-npm`: to generate a NOTICE file for npm dependencies
  called NOTICE-npm (using Licensed).

These user-facing commands rely on a couple of underlying Make targets
that will create the Licensed binary and initialize the Licensed cache
correctly.

The new commands are also used in the CI job for checking licenses to
simplify the job logic as well as avoid duplicate logic for running
Licensed.