Add markdownlint
Add markdownlint through the markdownlint-cli package and add a Make
target to run it on all the MarkDown files in the project - using the
configuration from the config file added (add for convenience, currently
it contains nothing what you'd consider configuration).

The markdownlint-cli package is an npm dependency. So, as usual, the
executable of the package is made available as an alias npm script and
invoked in the Makefile that way.

At the moment, markdownlint does not detect any issues with the MarkDown
files - which is good :)