Update reusable-audit.yml to take a list of refs as input
Instead of users defining a matrix over the workflow, the workflow now
takes a list of refs will build a matrix based on that for the relevant
jobs. This avoids the secrets job being repeated multiple times for no
reason, in particular when invoked on schedule.

Update usage of the reusable workflow accordingly.