Create initial script to check licenses in the SBOM
Write a script to evaluate the licenses found by Syft and put in the
SBOM. If any "violations" (defined below) are found it exists with a
non-zero exit code, first outputting all violations found. Else it exits
with the exit code 0.

A "violation" is defined as a project that does not have any license
that's in the approved lists of licenses (which is defined at the top of
the script). The current list is based purely on what licenses are
currently in the project's SBOM. Some notable inclusions:

- "BSD": not a standard license name - it appears in some sample code of
  an ISC licensed dependency of npm itself.
- "Apache-2.0" and "Apache 2.0": the latter is not standard, but one
  dependency uses this spelling for the former license. I think it's
  safe to assume any dependency that uses the latter intents to former.
- "OpenSSL" and "Zlib": project-specific licenses, based on my research
  of these licenses these are safe to allow.
- "GPL-2.0-only" and "GPL-2.0-or-later": just the name "GPL" may invoke
  fear in some. For this project I believe using GPL-2.0'ed dependencies
  is fine because the project isn't distributed in binary form.

This script is intended to evolve with the project's needs, this is
really just an initial, quick, iteration.