Add custom timeout for running tests
Define a custom timeout when running tests with AVA to ensure tests do
not fail due to timeouts - especially in continuous integration (CI)
jobs. The value of `20s` (20 seconds) was derived from successful CI
jobs, where it could take up to 9.9 seconds for (typically the first)
tests to finish.

The primary reason tests for this project can be slow is because each
test starts a docker container and waits for it to terminate.

The timeout being configured is per-test (not overall). On top of
increasing the timeout, this also makes the timeout explicit and
tracked in version control, making it easier to reference in the future.