Add Make command to validate licenses and run it in the CI
Create a new Make command to run the `check-licenses.js` script after,
if necessary, (re)generating the SBOM. Also add a continuous integration
(CI) job that runs the license check (using the Make command).

Temporarily disallows two licenses to demonstrate the behaviour of the
script and CI job if a disallowed license is detected.