Create Make command to check dependency licenses using Fossa
Create a new Make command that can be used to check the licenses of
dependencies using the Fossa CLI. The command assumes the Fossa CLI is
installed on the system, hence it's added as a prerequisites to the
Contributing Guidelines.

The continuous integration job for checking licenses is updated to use
this new command as well in order to avoid duplication.

The reason it's assumed to be on the user's system is because there's no
support for local installation of the Fossa CLI, so unlike e.g. Syft, it
can't easily be installed into the project's .bin/ directory.