Add Make command to validate npm licenses
Create a new Make command to check licenses, for now only npm licenses,
using "licensee". licensee is an npm dependency/tool to validate the
licenses of npm dependencies. Like with AVA, eslint, and markdownlint,
licensee is aliases as an npm script for convenient access in the
Makefile. The Makefile simply invokes licensee and tells it to be quiet
(otherwise it will output license details about all dependencies). This
way the `license-check` command exits with a non-zero exit code if there
are any license violations and a zero exit code otherwise.

An initial configuration file for licensee is included. This specifies
an allowlist of licenses - based on the licenses of current
dependencies. Additionally, the "corrections" field is present to
disable this feature in favor of manual review (enabling this makes
licensee cross reference against a public list of license corrections in
case a dependency didn't specify their license correctly).