Create make command to lint the Dockerfile
Create a new command for Make in the Makefile for linting the Dockerfile
using hadolint [1].

hadolint is not added as a separate dependency in any other way because
it's being run through Docker, which is already a project dependency. It
is added in a way similar to Syft and Grype though, with the version
that is being used defined at the top of the `Makefile`.

The new command is added to the list of `.PHONY` commands as well since
it's not tied to a file. For consistency the `lint-md` command has been
added as well (since it was missing).

--
1. https://github.com/hadolint/hadolint