Add hadolint configuration
Add a hadolint configuration file with an initial configuration and
update the `lint-docker` command to use this configuration file.

The configuration file is primarily aimed to align with the existing
Dockerfile (in particular w.r.t. to labels and the registry). The
configuration options for rules are present but unconfigured, just to
make it easier to change them if necessary.