Update release automation and guidelines
Following the failure of the last automated release, fix the validation
job of the release workflow by fetching submodules (otherwise the tests
fail).

Also, update the release guidelines to correctly state that the
continuous delivery "may kick in" after pushing the tag, rather than
after pushing the tag and `v0` branch. This is because the latter should
be carried out by the continuous delivery.

Lastly, align the name of the checkout steps in the release workflow
with the (targeted) naming format for this project's workflows.