Parse Licensed config file for licenses in check-licenses.js
Update the check-licenses.js script to get the list of allowed licenses
from the Licensed configuration file. Update the Licensed configuration
accordingly to include the license identifiers it was missing w.r.t. the
script.

As the Licensed configuration file is a YAML file, the script requires a
new npm dependency to parse YAML files. This was added in the form of
the package named "yaml" [1].

--
1. https://www.npmjs.com/package/yaml