Initialize dev docs with license validation decision record
Initialize a developer documentation directory at docs/dev/ with a
README explaining the idea and an initial decision record describing the
decision process behind the license validation tooling used by the
project.

The idea behind this documentation is to embed all context that
currently only exists on GitHub within the repository so it can be
preserved better.

The "decision record" name and format is inspired by the concept of
"Architecture Decision Records". However, as the scope of these docs is
broader than architecture, I've decided to just name it decision
records.