Try out `anchore/sbom-action` for a dependency snapshot of system deps
Add the `anchore/sbom-action` Action [1] to the "Audit / Docker" job
with the primary aim of trying out its `dependency-snapshot` options.
This option allows for uploading the dependency information found by
Syft [2] to be uploaded to GitHub's dependency graph [3].

Note that in this commit there may be some overlap in functionality
between the newly added Action and the pre-existing job steps. If true,
this is to be resolved later.

Also, temporarily run the reusable-audit.yml workflow from the current
branch in the "Check" workflow for testing purposes.

--
1. https://github.com/marketplace/actions/anchore-sbom-action
2. https://github.com/anchore/syft
3. https://docs.github.com/en/rest/dependency-graph/dependency-submission