Create decision record regarding testing
Create a second decision record on the topic of testing. It attempts to
capture everything that was considered when setting up testing. Since
the setup of testing was less involved than license validation, the
format is quite different.

The `docs/dev/README.md` is updated with a link to the new decision and
both this README and the license validation decision record have updated
comments for links in-line with the testing decision record.