Create decision record regarding dependency management
Create a second decision record on the topic of dependency management.
It attempts to capture everything that was considered when setting up
dependency management. Like testing, the setup of dependency management
was less involved than license validation, hence the format is more
akin to the testing decision record.

The `docs/dev/README.md` is updated with a link to the new decision.