Bump Licensed from 3.8.0 to 3.9.0
Update the version of Licensed used from 3.7.3 to 3.8.0. Use the new
`--computed` option on the "notice" command to avoid having to create.
Combined with the update to the "status" command in [1], the "cache"
command now no longer needs to be invoked to create a license cache.
Note that the Licensed cache directory is still used as that's where
Licensed puts the NOTICE file.

--
1. cc929e846bb3d8cb0ae189f054cd1acd504a86c6