Fix Licensed-related targets dependencies
Update both the "license-check-npm" and "notice-npm" targets to depend
on the node_modules/ directory (because Licensed expects it to be there,
prior to changes in [1, 2] this was handled by the `$(LICENSED_CACHE)`
target) and, in case of the latter target, on the .tmp/ directory.

The latter involves additionally changes. Recall that Licensed puts the
NOTICE file in the configured `cache_path`. This was fine prior to [1]
when this directory would be created when running `licensed cache`, but
now this folder is not created yet. Unfortunately, `licensed notice`
doesn't create this folder if it doesn't exist. To remediate this, the
`cache_path` is now just the .tmp/ directory and the "notice-npm" target
depends on the `$(TEMP_DIR)` target to ensure that directory is created.

--
1. d9e291573c447fd3cc7f9dee2e3c419dc43b0598
2. cc929e846bb3d8cb0ae189f054cd1acd504a86c6