Add test data and test snapshot for ignoreing test files
Create a new sample project as testdata for the purposes of testing
ignore patterns for test files. Vulnerable regular expression in tests
aren't interesting so shouldn't be reported by this scanner.

Note that this is partially covered by the _shescape_ test project as
well. The sample project was added to cover test approaches not covered
by _shescape_ (such as using ".spec.js")