Use and configure yamllint
Add a make target to run yamllint [1] and configure yamllint through the
`.yamllint.yml` configuration file. Also document the usage of yamllint
in the Contributing Guidelines and run it in the continuous integration
"Check / Lint" job.

Note, as of this commit there are still (preexisting) yamllint
violations in this project.

--
1. https://github.com/adrienverge/yamllint