Update release guidelines and release automation
Remove the no-longer-necessary step of updating test snapshots every
release (and, correspondingly, including such changes in the version
bump commit) from the release guidelines.

Update the release automation for updating the major version branch to
fetch the entire commit history in an effort to avoid errors when
pushing an update to the major version branch.