Remove "Licensed"
Remove the dependency on "Licensed" as a license validation/compliance
tool and disable the `license-check-npm` target. The `notice-npm` target
is removed entirely as it's underused and thus isn't planned to be
replaced.

The choice to remove Licensed stems from the stems from the fact that it
isn't able to recognize the licenses for (at)typescript-eslint's parser
and typescript-estree packages. Having to update these exceptions every
time they're updated is unnecessarily involved, especially considering
they update not-infrequently.

Note: the configuration file for Licensed is kept intentionally as the
license validation script for the SBOM generated for the Docker image,
`scripts/check-licenses.js`, uses it as it's source of allowed licenses.