Add "licensee" for `license-check-npm`
Add "licensee" as an npm dependency to be used for license checking of
npm dependencies. Add it as an `npm script` alias similar to other npm
CLI tooling. Use this alias in the Makefile for the `license-check-npm`
target - the `--errors-only` flag instructs licensee to limit its output
to only include dependencies licensed under a not-explicitly-allowed
license.

The `.licensee.json` file configures licensee. The configuration details
are based on the now-removed configuration for Licensed. Two additional
licenses are added to the allowlist, "CC0-1.0" and "CC-BY-3.0", which
are used by some transitive dependencies of licensee. The check-licenses
script has been updated to get the license allowlist from this file
instead of the Licensed config file (which allows for the removal of the
npm dependency `yaml`, which was used only for parsing the Licensed
configuration file).

Note that no exceptions are required for unrecognized licenses(!). This
is due to the `"corrections': true` configuration. Without it, one
dependency's (docopt) license isn't recognized. This options tells
licensee to use a crowdsourced list of license "corrections" (called so
because its scope is providing corrections for packages that don't
specify their license in an npm-standard way).