Set up cosign-based signing for releases to Docker hub
Update the release workflow to automatically sign and publish signatures
for the Docker image it produces to Docker hub using cosign [1]. To
achieve this, access to the `id-token` is provided (required for keyless
signing), cosign is installed using the `sigstore/cosign-installer`
(latest version as of this commit is used), and the created Docker image
is signed utilizing the outputs of the `docker/build-push-action`
action.

Note: as of this commit keyless signing as seen here is still
experimental!

--
1. https://github.com/sigstore/cosign