Use order-only dependency on .bin/ and .tmp/
Update the `$(TEMP_DIR)/dockerimage` target to have .tmp/ as an
order-only dependency in order to prevent rebuilding of the Docker image
whenever something in the .tmp/ dir changes (which inherently updates
the last-updated timestamp of .tmp/).

Similarly, update the `$(SYFT)` and `$(GRYPE)` to have an order-only
dependency on .bin/ - previously there was no dependency because these
targets work without .bin/ (they create it if necessary) and having a
"normal" dependency caused unnecessary rebuilds.