Add support for specifying tags with `make build`
Update the `Makefile` to allow for specifying a image tag when running
`make build`. The default tag is now explicitly "latest". The temporary
file corresponding to the image is now specific to the tag being used,
these files are grouped in the directory `.tmp/dockerimages` (plural as
it may contain multiple images and, crucially, `.tmp/dockerimage` is
likely to already exist as a file due to the previous setup).

Update the Release Guidelines steps related to building (and publishing)
accordingly to use the `make build` command over `docker build`. Also
fix a grammar error in the first line of step 11.