Remove duplicated metadata from `package.json`
Update the `package.json` file by removing the "name", "description",
and "license". All this information is present in the Dockerfile/image
as labels, so is duplicated at least from that perspective.
Additionally, the "name" and "description" are present in the README.md
and the "license" is present in the README.md and LICENSE file.

Also include a correction to the Release Guidelines, removing an unused
link reference definition.