Fix cosign usage in "Release / Docker Hub" job
Update the "Release / Docker Hub" job to correct the usage of cosign [1]
for keyless signing [2] of the Docker image published to Docker Hub [3].
As an experimental feature, keyless signing requires the environment
variable `COSIGN_EXPERIMENTAL` to be set. Additionally, since it is an
experimental feature, it's usage might break on non-major version
releases of cosign. Thus, cosign is pinned to the current latest
version.

--
1. https://github.com/sigstore/cosign
2. https://github.com/sigstore/cosign/blob/44d58de/KEYLESS.md
3. https://hub.docker.com/repository/docker/ericornelissen/js-re-scan