Add support for corrections to SBOM license checker
Update the check-licenses.js script to support corrections. This
supports the use case where Syft [1] was unable to determine the license
of a dependency.

Corrections are identified by the dependencies name and must include the
actual license(s), including a URL where the license(s) are specified,
and a reason for the correction.

--
1. https://github.com/anchore/syft