Track tool versions in `.tool-versions`
Create a `.tool-versions` file with a select few tools and their
corresponding version. The file name and  format is based on asdf [1]
but is also human-readable so suitable for simple reference as well.
Having this available should help with reproducability and help
contributors figure out how to get the correct outputs for each of these
tools.

Also use asdf in the CI, where applicable, to install the correct
versions of these tools. This enhances the reproducibility of the CI
jobs.

Notably `yamllint` is absent from the file. This is because there does
not appear to be a asdf plugin for yamllint. Also, Syft and Grype aren't
included (for now) because their versions are tracked in the `Makefile`.

--
1. https://asdf-vm.com/