Create `make verify` command
Update the Makefile to include a phony target called 'verify' that can
be used to "Verify project is in a good state". This is a convenience
target/command to quickly run tests and other checks.

Use this command to verify the state of the project prior to releasing
a new version in the release automation.