Track Syft & Grype versions in `.tool-versions`
Move the recorded versions for both Grype [1] and Syft [2] from the
`Makefile` to the `.tool-versions` file. This helps consolidate tooling
version information in a single place and allows for installing the
preferred versions of these tools using asdf [3].

As a consequence, the manual installation logic for Grype and Syft has
been removed. All related Make targets were updated accordingly.

The CI workflows have been updated to install tooling using the Action
`asdf-vm/actions/install` as necessary. Note that this is a pre-existing
Actions dependency (see `check.yml`'s job named "Lint").

The Contributing Guidelines now also include Grype and Syft as optional
prerequisites since they're no longer installed as needed but instead
need to be made available on the system manually (even if that's using
asdf [3]). An additional improvement to the Contributing Guidelines is a
reference to `.tool-versions` for version information for tools covered
in the `.tool-versions` file.

--
1. https://github.com/anchore/grype
2. https://github.com/anchore/syft
3. https://asdf-vm.com/