Update tooling nightly
Add a new job to the `nightly.yml` workflow to try to update tooling
(defined in `.tool-versions`) nightly. It will always try to update to
the latest available version. If and only if a new version is available
a Pull Request will be created with the changes to update the tool. This
should help ensure tooling is kept up-to-date.

For now, the list of tools is duplicated between the `.tool-versions`
file and this workflow (see the `matrix`). This may be improved upon in
the future.