Track yamllint versions in `.tool-versions`
Add the used version of yamllint [1] to the `.tool-versions` file. This
helps consolidate tooling version information in a single place and
allows for installing the preferred versions of yamllint using asdf [2].

The nightly tooling update job as well as the Contributing Guidelines
have been updated in accordance with this addition to `.tool-versions`.

--
1. https://github.com/adrienverge/yamllint
2. https://asdf-vm.com/