Auto-install tooling with asdf
Update Make targets requiring tooling that is installed by asdf [1] to
depend on a file that's created after `asdf install` is run [^1]. Since
asdf isn't a requirement for this project, it's called conditionally if
available on the system. Else, the tool won't be installed live and
instead a system-level instance of the tool will be used (if available).

The dependency on `.tool-versions` ensures the available version of all
tools always matches what's configured in the repository.

--
^1: This approach is similar to what's used for Docker images with files
    in `.tmp/dockerimages/`.

--
1. https://asdf-vm.com/