Disable `npm audit` when install dependencies inside the Docker image
Update the Dockerfile to configure npm to **not** audit dependencies
when installing dependencies when building the Docker image. When the
image is being build this is not relevant information (this project has
other, explicit, ways to keep an eye on known vulnerabilities in third-
party components) and only really clutters the build logs.