Create automated workflow for initiating a new release
Add a new GitHub Actions workflow to initiate the release process by
automatically performing steps 1 to 5 of the (manual) release steps. The
workflow is manually triggered to give maintainers full control over
when a new release is initiated.

The Release Guidelines are updated to recommend the use of this
automated approach.