Create test sample for custom ignore patterns
Add a new sample project to the testdata/ which is intended to verify
that ignore patterns provided by the user through `--ignore-pattern` are
respected by the scanner.

This is the first test with custom CLI arguments. To achieve test case
specific arguments a new file is introduced called `.args` which defines
the arguments for the test. The test case will read this file and add
the arguments defined in the file to the scanner invocation. This way
the resulting snapshot will be for a scan with the defined arguments.

For backwards (test) compatibility, and to avoid having to create this
file unnecessarily, the missing-`.args` file is handled by simply not
using any additional arguments.