Add rtx as automatic .tool-versions installer in Makefile
Update the Makefile to extend the automatic tooling installation support
to use rtx if available. The Makefile was refactored to make it more
explicit tooling isn't managed solely through asdf.

This also includes a small improvement to this functionality where it
won't keep trying to install tools if asdf (or now rtx) isn't available
on the system.