Track cosign version in .tool-versions
Add cosign to the .tool-versions file to track it alongside other tools.
This allows it to be incorporated in the nightly tooling update workflow
- since the asdf cosign plugin is hosted on GitLab, gitlab.com as added
as one of the `allowed-endpoints`. The workflow that uses cosign is
updated to use the version of cosign specified in .tool-versions.

Also, include a minor change to the nightly tooling workflow to use a
branch name that is tooling-version-manager independent.