Update cosign command to skip prompt
Based on the cosign-installer action's [1] v3.0.1 documentation, update
the usage of `cosign sign` to include the `--yes` flag. This _should_
avoid being prompted for confirmation on using keyless signing, which
caused the continuous deployment of v0.4.2 to fail.

In particular, this is based on the following snippet (modified for
brevity):

      - name: Sign the images with GitHub OIDC Token
        run: cosign sign --yes ${TAGS}
        env:
          TAGS: ${{ steps.docker_meta.outputs.tags }}

--
1. https://github.com/marketplace/actions/cosign-installer