Set up Prettier for JavaScript linting
Add Prettier [1] as a devDependency and use it for linting JavaScript
files, in particular to enforce a consistent style across the project.
A `Makefile` target is provided to lint just JavaScript files, this is
also where the configuration of Prettier lives (this may be re-evaluated
in the future). The `make lint` target is updated to also lint `.js`
files. The "Check / Lint" job has been updated to lint JavaScript files.
Finally, the Contributing Guidelines have been updated to cover linting
of JavaScript files as well.

JavaScript files in the `testdata/` directory aren't linter because 1)
their style is less relevant as they're not meant to be read or edited
(regularly), and 2) this can get messy with submodules.

--
1. https://prettier.io/